% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.features.list.R
\name{conv.features.list}
\alias{conv.features.list}
\title{Create converted feature matrix using a list of list of ligands and a conversion matrix
Uses conv.features()}
\usage{
conv.features.list(ligands.list, conv.table, n)
}
\arguments{
\item{ligands.list}{List of list of HLA allele ligand sequences created by read.ligands().}

\item{conv.table}{Feature conversion matrix created by read.AA.table().}

\item{n}{Length of peptides in ligands to convert into features.}
}
\description{
\code{conv.features.list()} takes in a file directory and returns a list of lists of each HLA's ligands
}
