% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.features.R
\name{conv.features}
\alias{conv.features}
\title{Create converted feature matrix using a list of ligands and a conversion table}
\usage{
conv.features(ligands, conv.table, n)
}
\arguments{
\item{ligands}{List of HLA ligand sequences from one HLA allele.}

\item{conv.table}{Feature conversion table created by make.conv.table().}

\item{n}{Length of peptides in ligands to convert into features.}
}
\description{
\code{conv.features()} takes in a list of ligands from one HLA and returns
a converted feature matrix of the HLA's ligands using a conversion matrix.
}
