% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptidePCA.R
\docType{data}
\name{convmat.24}
\alias{convmat.24}
\title{Amino acid conversion matrix with 24 properties}
\format{A data frame with 20 rows (amino acids) and 24 columns (properties)}
\usage{
data(convmat.24)
}
\description{
Conversion matrix used to transform n-mer ligand sequences into 24 * n position-by-property features.
4 properties are physicochemical: molecular weight, surface area, hydrophobicity index, and isoelectric point.
20 properties describe identity of the amino acid at a position.
}
\references{
\url{to do}
}
\keyword{datasets}
