% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptidePCA.R
\docType{data}
\name{convmat.4}
\alias{convmat.4}
\title{Amino acid conversion matrix with 4 properties}
\format{A data frame with 20 rows (amino acids) and 4 columns (properties)}
\usage{
data(convmat.4)
}
\description{
Conversion matrix used to transform n-mer ligand sequences into 4 * n position-by-property features.
4 properties are physicochemical: molecular weight, surface area, hydrophobicity index, and isoelectric point.
}
\keyword{datasets}
