% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptidePCA.R
\docType{package}
\name{LAPP}
\alias{LAPP}
\alias{LAPP-package}
\title{PeptidePCA}
\description{
PeptidePCA contains a small suite of functions that can be used to recreate PCA figures from Hilton et al. (2017).
HLA peptide sequences can be transformed into new features using an amino acid conversion matrix and then plotted
using PCA. The packages ggplot2 and FactoMineR are required.
}
