% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptidePCA.R
\docType{data}
\name{ligands.4A}
\alias{ligands.4A}
\title{Ligand sequences from 4 HLA-A alleles}
\format{A list of 4 character lists}
\usage{
data(ligands.4A)
}
\description{
List of list of sequences from HLA-A*01:01, HLA-A*02:01, HLA-A*03:01, and HLA-A*11:01.
All sequences from obtained from the publically shared sequences from IEDB on November 2016.
}
\keyword{datasets}
