% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.conv.table.R
\name{make.conv.table}
\alias{make.conv.table}
\title{Read in amino acid table}
\usage{
make.conv.table(conv.table.file)
}
\arguments{
\item{AA.file}{csv file to be converted into the amino acid conversion matrix. See given example files for the format of these input files.}
}
\description{
\code{make.conv.table()} takes in a csv file to create a matrix of amino acid feature to be used in conversion
}
