% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep.pca.R
\name{pep.pca}
\alias{pep.pca}
\title{Perform PCA on HLA features}
\usage{
pep.pca(features)
}
\arguments{
\item{features}{data.frame object returned by make.features.list().}
}
\description{
\code{pep.pca()} takes in a data.frame from conv.features.list() and performs PCA.
The results are a modified form of results from PCA() from the FactoMineR package.
}
