% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ligands.R
\name{read.ligands}
\alias{read.ligands}
\title{Read in all ligand files in a folder}
\usage{
read.ligands(lig.dir, lig.names = NA)
}
\arguments{
\item{lig.dir}{file directory with one or more files of txt files of peptide sequences.
Each file is named after the HLA allele, and one peptide sequence is present per line.}

\item{lig.names}{character list of HLA names for the ligand files. If not supplied, the names
of the .txt files will be used.}
}
\description{
\code{read.ligands()} takes in a file directory and returns a list of lists of each HLA's ligands
}
